replace cli with oclif
