+++ src/bin/run
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const Command = require('..')
+const Command = require('../command')
 
 Command
   .run()
